Fixed 2 step in channel creation
